+++ src/main/java/lucie/deathtaxes/registry/AttachmentTypeRegistry.java
@@ -1,5 +1,6 @@
 package lucie.deathtaxes.registry;
 
+import com.mojang.serialization.Codec;
 import lucie.deathtaxes.DeathTaxes;
 import net.minecraft.world.item.component.ItemContainerContents;
 import net.neoforged.neoforge.attachment.AttachmentType;
@@ -14,4 +15,7 @@ public class AttachmentTypeRegistry
 
     public static final Supplier<AttachmentType<ItemContainerContents>> PLAYER_INVENTORY_DROPS = ATTACHMENT_TYPES.register("player_inventory_drops", () ->
             AttachmentType.builder(() -> ItemContainerContents.EMPTY).serialize(ItemContainerContents.CODEC).build());
+
+    public static final Supplier<AttachmentType<Long>> DESPAWN_TIME = ATTACHMENT_TYPES.register("despawn_time", () ->
+            AttachmentType.builder(() -> 0L).serialize(Codec.LONG).build());
 }